rename new wrapper to old name, remove EPP dependency
