+++ src/XlsToEfCore/Import/ExcelIoWrapper.cs
@@ -1,22 +1,19 @@
 ﻿using System.Collections.Generic;
+using System.Data;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
-using OfficeOpenXml;
+using ExcelDataReader;
 
 namespace XlsToEfCore.Import
 {
-    public interface IExcelIoWrapper
-    {
-        Task<IList<string>> GetSheets(string filePath, FileFormat fileFormat);
-        Task<IList<string>> GetSheets(Stream fileStream, FileFormat fileFormat);
-        Task<List<Dictionary<string, string>>> GetRows(string filePath, string sheetName, FileFormat fileFormat);
-        Task<List<Dictionary<string, string>>> GetRows(Stream fileStream, string sheetName, FileFormat fileFormat);
-        Task<IList<string>> GetImportColumnData(XlsxColumnMatcherQuery matcherQuery);
-    }
-
     public class ExcelIoWrapper : IExcelIoWrapper
     {
+        public ExcelIoWrapper()
+        {
+            System.Text.Encoding.RegisterProvider(System.Text.CodePagesEncodingProvider.Instance);
+        }
+
         public async Task<IList<string>> GetSheets(string filePath, FileFormat fileFormat)
         {
             using (var stream = new FileInfo(filePath).OpenRead())
@@ -24,100 +21,135 @@ namespace XlsToEfCore.Import
                 return await GetSheets(stream, fileFormat);
             }
         }
-
-        public async Task<IList<string>> GetSheets(Stream fileStream, FileFormat fileFormat)
+        
+        public Task<IList<string>> GetSheets(Stream fileStream, FileFormat fileFormat)
         {
-            var sheetNames = await Task.Run(() =>
-            {
-                using (var excel = new ExcelPackage(fileStream))
-                {
-                    return excel.Workbook.Worksheets.Select(x => x.Name).ToList();
-                }
-            });
-
-            return sheetNames;
-
+            var reader = GetReader(fileStream, fileFormat);
+            var worksheetNames = GetAllSheetNames(reader);
+            return Task.FromResult<IList<string>>(worksheetNames);
         }
 
-        private async Task<IList<string>> GetColumns(string filePath, string sheetName)
+        public async Task<List<Dictionary<string, string>>> GetRows(string filePath, string sheetName, FileFormat fileFormat)
         {
             using (var stream = new FileInfo(filePath).OpenRead())
             {
-                return await GetColumns(stream, sheetName);
+                return await GetRows(stream, sheetName, fileFormat);
             }
         }
 
-        private async Task<IList<string>> GetColumns(Stream fileStream, string sheetName)
+        public Task<List<Dictionary<string, string>>> GetRows(Stream fileStream, string sheetName,
+            FileFormat fileFormat)
         {
-            var colNames = await Task.Run(() =>
+            var reader = GetReader(fileStream, fileFormat);
+
+            if (fileFormat == FileFormat.OpenExcel)
             {
-                using (var excel = new ExcelPackage(fileStream))
+                EnsureSheetExists(sheetName, reader);
+                SelectSheetOnReader(sheetName, reader);
+            }
+
+            var columns = GetColumnsFromPreSelectedSheet(reader);
+            var rows = new List<Dictionary<string, string>>();
+
+            while (reader.Read())
+            {
+                var row = new Dictionary<string, string>();
+                var record = (IDataRecord) reader;
+                for (var field = 0; field < record.FieldCount; field++)
                 {
-                    EnsureSheetExists(sheetName, excel);
-
-                    var sheet = excel.Workbook.Worksheets.First(x => x.Name == sheetName);
-                    var headerCells =
-                        sheet.Cells[
-                            sheet.Dimension.Start.Row, sheet.Dimension.Start.Column, 1, sheet.Dimension.End.Column];
-                    return headerCells.Select(x => x.Text).ToList();
+                    row.Add(columns[field], record[field].ToString());
                 }
-            });
 
-            return colNames;
+                rows.Add(row);
+            }
+
+            return Task.FromResult(rows);
         }
 
-        private static void EnsureSheetExists(string sheetName, ExcelPackage excel)
+        private static IExcelDataReader GetReader(Stream fileStream, FileFormat matcherQueryFileFormat)
         {
-            if (excel.Workbook.Worksheets.All(x => x.Name != sheetName))
+            if (matcherQueryFileFormat == FileFormat.OpenExcel)
+                return ExcelReaderFactory.CreateOpenXmlReader(fileStream);
+            else return ExcelReaderFactory.CreateCsvReader(fileStream);
+        }
+
+        private static void SelectSheetOnReader(string sheetName, IExcelDataReader reader)
+        {
+            for (var i = 0; i < reader.ResultsCount; i++)
+            {
+                if (reader.Name == sheetName)
+                    break;
+                reader.NextResult();
+            }
+        }
+
+        private static List<string> GetAllSheetNames(IExcelDataReader reader)
+        {
+            var worksheetNames = new List<string>();
+            for (var i = 0; i < reader.ResultsCount; i++)
+            {
+                worksheetNames.Add(reader.Name);
+                reader.NextResult();
+            }
+            reader.Reset();
+            return worksheetNames;
+        }
+
+        private static void EnsureSheetExists(string sheetName, IExcelDataReader excel)
+        {
+            if (GetAllSheetNames(excel).All(x => x != sheetName))
                 throw new SheetNotFoundException(sheetName);
         }
 
         public Task<IList<string>> GetImportColumnData(XlsxColumnMatcherQuery matcherQuery)
         {
-            if(matcherQuery.FileStream == null)
-                return GetColumns(matcherQuery.FilePath, matcherQuery.Sheet);
+            if (matcherQuery.FileFormat == null)
+                matcherQuery.FileFormat = FileFormat.OpenExcel;
+            if (matcherQuery.FileStream == null)
+                return GetColumns(matcherQuery.FilePath, matcherQuery.Sheet, matcherQuery.FileFormat.Value);
 
-            return GetColumns(matcherQuery.FileStream, matcherQuery.Sheet);
+            return GetColumns(matcherQuery.FileStream, matcherQuery.Sheet, matcherQuery.FileFormat.Value);
         }
 
-        public async Task<List<Dictionary<string, string>>> GetRows(string filePath, string sheetName,
-            FileFormat fileFormat)
+        private async Task<IList<string>> GetColumns(string filePath, string sheetName, FileFormat fileFormat)
         {
             using (var stream = new FileInfo(filePath).OpenRead())
             {
-                return await GetRows(stream, sheetName, FileFormat.OpenExcel);
+                return await GetColumns(stream, sheetName, fileFormat);
             }
         }
 
-        public async Task<List<Dictionary<string, string>>> GetRows(Stream fileStream, string sheetName, FileFormat fileFormat)
+        private async Task<IList<string>> GetColumns(Stream fileStream, string sheetName,
+            FileFormat fileFormat)
         {
-            var worksheetRows = await Task.Run(() =>
+            var colNames = await Task.Run(() =>
             {
-                using (var excel = new ExcelPackage(fileStream))
+                var reader = GetReader(fileStream, fileFormat);
+
+                if (fileFormat == FileFormat.OpenExcel)
                 {
-                    EnsureSheetExists(sheetName, excel);
-
-                    var sheet = excel.Workbook.Worksheets.First(x => x.Name == sheetName);
-
-                    var rows = new List<Dictionary<string, string>>();
-
-                    for (var rowNum = 2; rowNum <= sheet.Dimension.End.Row; rowNum++)
-                    {
-                        var rowDict = new Dictionary<string, string>();
-                        var row = sheet.Cells[string.Format("{0}:{0}", rowNum)];
-
-                        for (int colIndex = sheet.Dimension.Start.Column; colIndex <= sheet.Dimension.End.Column; colIndex++)
-                        { // ... Cell by cell...
-                            string cellValue = sheet.Cells[rowNum, colIndex].Text; // This got me the actual value I needed.
-                            rowDict.Add(sheet.Cells[1, colIndex].Text, cellValue);
-                        }
-                        rows.Add(rowDict);
-                    }
-
-                    return rows;
+                    EnsureSheetExists(sheetName, reader);
+                    SelectSheetOnReader(sheetName, reader);
                 }
+
+                return GetColumnsFromPreSelectedSheet(reader);
             });
-            return worksheetRows;
+
+            return colNames;
+        }
+
+        private static List<string> GetColumnsFromPreSelectedSheet(IDataReader reader)
+        {
+            var headerNames = new List<string>();
+
+            reader.Read();
+            var record = (IDataRecord) reader;
+            for (var field = 0; field < record.FieldCount; field++)
+            {
+                headerNames.Add(record[field].ToString());
+            }
+
+            return headerNames;
         }
     }
 }
